etude_dir {etude}	R Documentation

Make a listing of an etude directory

Description

Etude files are stored in a directory, typically "Exercises". This function scans through the RMD files in such a directory and constructs a data frame based on the YAML. The data frame can be useful, for instance, for including in a parent file using etude_list().

Usage

etude_dir(where)

Arguments

where
the name of the folder containing the etude RMD files

[Package etude version 0.3.0.9001 Index]