etude_list {etude}	R Documentation

Include multiple etude exercises

Description

Include multiple etude exercises

Usage

etude_list(spec, keep_ref = FALSE)

Arguments

spec
A data frame with one row for each exercise to be inclued in the parent document. The variables must be named exercise, answers, documentation, package, and name. (Any additional variables will simply be ignored by etude_list.)

keep_ref
A logical flag, FALSE by default. If TRUE then lines defining pandoc references will be retained in the output document. If the parent document is being compiled using bookdown, this is what you want, since the references will be replaced appropriately by pandoc. But if not using bookdown, the references will appear verbatim in the output, which is probably not what you want.

Examples

## Not run: 
etude_list(
 tibble(exercise="dog-drink-ring",
      answers=FALSE, documentation=FALSE, name="Problem 1: "
      )
)

## End(Not run)

[Package etude version 0.3.0.9001 Index]