etude_xform {etude}	R Documentation

Insert an exercise from a file

Description

Each exercise is contained in its own file, with a unique ID. This function reads the file for that ID, knits the contents, and returns the knitted contents. Typically this will be used to insert the exercise into an Rmd file. If not specified, argument show_answer will be read from options("show_answer").

Usage

etude_xform(
  id,
  show_answer = getOption("show_exercise", TRUE),
  verbose = TRUE,
  keep_ref = FALSE,
  prob_name = "Problem XX",
  package = "etude",
  format = ifelse(knitr::is_latex_output(), "latex", "html")
)

Arguments

id
the unique ID of the exercise file. Don't include the file type suffix ".Rmd" in the id.

show_answer
if TRUE, include the answer comments from the file.

verbose
Include the YAML information

keep_ref
Logical flag indicating whether to keep the bookdown text reference (that is, (ref:...)). Default: FALSE

prob_name
Character string naming the problem, e.g. "Prob 1.3"

package
character string naming the package from which the problem is to be taken

Details

When output is in html, use the answer-fragment class in CSS to format the answers.

Examples

## Not run: 
the_testing_dir <- system.file("Test_exercises", package = "SDSdata")
include_exercise("beech-run-mug", directory = the_testing_dir, format = "latex")

## End(Not run)

[Package etude version 0.3.0.9001 Index]