exercise_hook {etude}	R Documentation

Exercise interface to knitr

Description

A hook to include exercises in a document

Usage

exercise_hook(before, options, envir)

Arguments

before
As with all knitr hooks, indicates whether we are before or after the chunk is evaluated.

options
As with all knitr hooks, the options given to the chunk

envir
As with all knitr hooks, the environment in which the chunk will be evaluated.

Details

In the startup to the document, set knitr::knit_hooks$set(exercise = exercise_hook)

You don't call the hook directly. Instead, it is called by knitr whenever the chunk option exercise = <file_hash> is given. You should communicate using the chunk options, not the ar The chunk relevant options are:

exercise=: hash for exercise file

name=: the identifier for the exercise to print

answers=: TRUE or FALSE: show the answers?

package=: the package the exercise comes from.

Examples

## Not run: 
# Chunk options should be like this
exercise = "dog-flies-home", name="Exercise 36", answers = FALSE,

## End(Not run)

[Package etude version 0.3.0.9001 Index]