new_etude_learnr {etude}	R Documentation

Addin to make a new etude exercise

Description

Creates the boilerplate for an etude exercise. Typically, this will be called by the RStudio addin, but you can also run it directly when save=TRUE

Usage

new_etude_learnr(directory = ".")

new_etude(directory = ".", learnr = FALSE)

new_etude_template(save = TRUE, exercise_id = NULL, learnr = FALSE)

Arguments

directory
Path to the directory where the files go

learnr
Logical flag (default: FALSE) whether to use a learnr compatible document for the exercise.

save
Logical indicating whether to save the template to a new file and return the name of that file

exercise_id
Root name for the exercise file (.Rmd will be added). By default, a random, unique root name will be created using the etude naming convention, e.g., bear-walk-dish

[Package etude version 0.3.0.9001 Index]